requirements-new.txt
scandir==1.6
h5py
scikit-image
dlib
face_recognition
youtube-dl
moviepy==1.0.0
numpy
tqdm
pydot
graphviz
tensorflow-gpu==2.4.0
dictionary
tqdm
pyfakewebcam
opencv-python
tf2onnx
protobuf==3.20.1